+++ apio-consumer/src/main/kotlin/com/liferay/apio/consumer/request/ApioCallback.kt
@@ -12,15 +12,26 @@
  * details.
  */
 
-package com.liferay.apio.consumer.configuration
+package com.liferay.apio.consumer.request
+
+import android.os.Build
+import android.support.annotation.RequiresApi
+import com.github.kittinunf.result.Result
+import com.liferay.apio.consumer.model.Thing
+import java.util.function.Consumer
 
 /**
  * @author Paulo Cruz
  */
-class RequestConfiguration @JvmOverloads constructor(
-	val embedded: List<String> = emptyList(),
-	val fields: Map<String, List<String>> = emptyMap(),
-	val headers: List<RequestHeader> = emptyList()) {
+@RequiresApi(Build.VERSION_CODES.N)
+class ApioCallback(private val onSuccess: Consumer<Thing>, private val onFailure: Consumer<Exception>)
+	: Function1<Result<Thing, Exception>, Unit> {
 
-	constructor(vararg headers: RequestHeader? = emptyArray()) : this(emptyList(), emptyMap(), headers.filterNotNull())
+	override fun invoke(result: Result<Thing, Exception>) =
+		result.fold({
+			onSuccess.accept(it)
+		}, {
+			onFailure.accept(it)
+		})
+
 }
